+++ src/main/java/com/comunidad/api/modelo/dao/interfaz/IEpsDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.comunidad.api.modelo.entidad.Eps;
 
-public interface IEpsDao extends CrudRepository<Eps, Long> {
+public interface IEpsDao extends CrudRepository<Eps, Integer> {
 
 }
