+++ src/main/java/com/comunidad/api/modelo/dao/interfaz/IModalidadDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.comunidad.api.modelo.entidad.Modalidad;
 
-public interface IModalidadDao extends CrudRepository<Modalidad, Long> {
+public interface IModalidadDao extends CrudRepository<Modalidad, Integer> {
 
 }
