+++ src/main/java/com/comunidad/api/modelo/dao/interfaz/IRolDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.comunidad.api.modelo.entidad.Rol;
 
-public interface IRolDao extends CrudRepository<Rol, Long> {
+public interface IRolDao extends CrudRepository<Rol, Integer> {
 
 }
