+++ src/main/java/com/comunidad/api/modelo/dao/interfaz/IUsuarioDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.comunidad.api.modelo.entidad.Usuario;
 
-public interface IUsuarioDao extends CrudRepository<Usuario, Long> {
+public interface IUsuarioDao extends CrudRepository<Usuario, Integer> {
 
 }
