+++ src/main/java/com/comunidad/api/modelo/dao/interfaz/IVisitanteDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.comunidad.api.modelo.entidad.Visitante;
 
-public interface IVisitanteDao extends CrudRepository<Visitante, Long> {
+public interface IVisitanteDao extends CrudRepository<Visitante, Integer> {
 
 }
