+++ src/main/java/com/comunidad/api/modelo/entidad/Modalidad.java
@@ -19,7 +19,7 @@ public class Modalidad implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String descripcion;
 	@OneToMany(mappedBy = "modalidad", cascade = CascadeType.REMOVE)
 	@JsonIgnoreProperties(value = { "modalidad" }, allowSetters = true)
@@ -30,7 +30,7 @@ public class Modalidad implements Serializable {
 		super();
 	}
 
-	public Modalidad(Long id, String descripcion, Set<Basico> basico) {
+	public Modalidad(Integer id, String descripcion, Set<Basico> basico) {
 		super();
 		this.id = id;
 		this.descripcion = descripcion;
@@ -42,11 +42,11 @@ public class Modalidad implements Serializable {
 		return "Modalidad [id=" + id + ", descripcion=" + descripcion + ", basico=" + basico + "]";
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
