+++ src/main/java/com/comunidad/api/modelo/servicio/implementacion/RolServicioImpl.java
@@ -11,11 +11,11 @@ import com.comunidad.api.modelo.entidad.Rol;
 import com.comunidad.api.modelo.servicio.interfaz.IServicio;
 
 @Service("RolServicioImpl")
-public class RolServicioImpl implements IServicio<Rol, Long> {
+public class RolServicioImpl implements IServicio<Rol, Integer> {
 
 	@Autowired
 	private IRolDao rolDao;
-	
+
 	@Override
 	@Transactional(readOnly = true)
 	public List<Rol> findAll() {
@@ -30,12 +30,12 @@ public class RolServicioImpl implements IServicio<Rol, Long> {
 
 	@Override
 	@Transactional(readOnly = true)
-	public Rol findById(Long id) {
+	public Rol findById(Integer id) {
 		return this.rolDao.findById(id).orElse(null);
 	}
 
 	@Override
-	public void deleteById(Long id) {
+	public void deleteById(Integer id) {
 		this.rolDao.deleteById(id);
 	}
